Alpha particles creation working,BreakoutFacade finished
Links between alpha and ball not visible because of a bug:
ball.getalphas(array) in breakoutfacade is always empty(no ball ever has
a link with an alpha...). I think this has something to do with cloning
and not. Internal adresses get lost when constantly updating the alphas
array in breakoutstate.

Newly added in breakoutstate: GetAlphas() & Concatenate() and Creation
of alpha particles in collideBallPaddle().